corrected threads randomly crashing; minor fixes
Failure to create pipe object returns `1` instead of `false` in all
`exec_internal()` calls (in fs, os, threads libraries).

Fixed the futures and `exec_internal()` function in threads.cpp - it was
crashing randomly (with error 32512) because, as it seems, `popen()` and
`pclose()` are not thread safe and hence cannot be executed
simultaneously. Adding a mutex/lock for `popen()` and `pclose()` fixed
that issue.