added src_idx (source index) in var_base_t
Addition of `src_idx` allows correct resolution in, say, `declared at
another location` error, allowing us to know in which source file,
exactly, was the variable created and using `parse_ctr` we can find the
location of the variable definition in that file.

In .gitignore, changed `tmp.et` to `tmp*.et`.

Version is now 0.1.1